pyproject.toml
[tool.poetry]
name = "click-skeleton"
version = "0.20"
description = "Click app skeleton"
authors = ["Adrien Pensart"]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
click-completion = "^0.5.0"
click-option-group = "^0.5.1"
click-aliases = "^1.0.1"
semver = "^2.10.2"
requests = "^2.24.0"
munch = "^2.5.0"
types-requests = "^2.25.0"
click = "^8.0.1"
click-didyoumean = "^0.3.0"
rich-click = "^1.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
coverage-badge = "^1.0.1"
restructuredtext-lint = "^1.3.2"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
mypy = "^0.931"
pytest-click = "^1.0.2"
mdformat = "^0.7.13"
isort = "^5.10.1"
bandit = "^1.7.2"
safety = "^1.10.3"

[tool.isort]
multi_line_output = 3

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov-report term-missing:skip-covered --cov-report html --durations=0 --cov click_skeleton -s"

[tool.pylint.master]
jobs = 4
[tool.pylint.messages_control]
disable = """
disable=too-many-instance-attributes,too-few-public-methods,logging-format-interpolation,line-too-long,too-many-arguments,protected-access,too-many-public-methods,too-many-instance-attributes,duplicate-code,too-many-nested-blocks,too-many-branches,too-many-return-statements,too-many-statements,too-many-locals,too-few-public-methods,too-many-ancestors,import-outside-toplevel,c-extension-no-member,logging-fstring-interpolation
"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
